docs: add details documentation of run calculation
